fix: Bug where view_functions were getting dropped
Fixes a bug that prevented the same URL from being registered with
different methods. If trying to register `POST /user` and `GET /user`,
one would get dropped because the internal mapping uses a dict. Changes
the key for that dict to be a namedtuple consisting of the rule and its
methods instead of just the rule itself.